payment type db tests passing, foreign key constraints on product tests

+++ bangazon-cli/Models/PaymentType.cs
@@ -5,6 +5,6 @@ namespace bangazon_cli.Models
         public int Id { get; set; }
         public int CustomerId { get; set; }
         public string Type { get; set; }
-        public long AccountNumber { get; set; }
+        public double AccountNumber { get; set; }
     }
 }
